Split code of CCF plugin in seperate components
